+++ CHANGELOG.md
@@ -47,6 +47,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - Quick range day filters now start from beginning of the day. remp/remp#605
 - Added `FORCE_HTTPS` environment variable to enforce HTTPS generation to URLs instead of determining protocol based on the request. This is useful in case you're running your application on `https`, but internally use proxy forwarding the request via `http`. remp/remp#619
 - Fixed add new ab variant replaces last variant instead of adding new after last one. remp/remp#634
+- Added option to disable banner events tracking. remp/remp#636
+- Added ability to access banner properties in custom JS code run in banner via newly added `params` object. remp/remp#636
 - Changed wording of hints in campaign's segment selection form. remp/remp#645
 - Fixed remplib initialization which could use misidentification of user - `remplib.getUserId()` would return `null` even when the `userId` was correctly set in `rempConfig`. remp/remp#651